autoconf: don't disable the autoheader warnings
The usual autoheader warning is due to AC_DEFINE variables not having a
description.  This results in no variable being defined in config.h, which leads
to code behaving as if the test failed when it actually succeeded.

This patch was introduced to OpenEmbedded back in 2004:

http://git.openembedded.org/openembedded/commit/?id=5eab06d132cb2895fd579f5cedffbb54c27794f8

There is no rationale for the patch so I suspect this is due to the warnings
being fatal and the submitter not understanding that the problem is more than
cosmetic.

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>